% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Analysis_Functions.R
\name{Logistic_Model}
\alias{Logistic_Model}
\title{A function}
\usage{
Logistic_Model(i, EXP, MET, Vars, Diag, SampleID)
}
\arguments{
\item{i}{Row Number to use in expression data}

\item{EXP}{- Expression data frame object colnames must equal MET rownames}

\item{MET}{- MetaData data frame object rownames must equal EXP colnames}

\item{Vars}{- Character vector of metadata columns to use in the model eg c("diagnosis", "msex", "APOE", "age_death", "pmi")}

\item{Diag}{- Character vector of metadata column to use as response variable must be binary factor eg. "diagnosis"}

\item{SampleID}{- Character vector of Meta Column name to use as sample ID eg. 'batchChannel'}
}
\value{
A vector
}
\description{
A function
}
