% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Analysis_Functions.R
\name{NeuroPath_Calc}
\alias{NeuroPath_Calc}
\title{A function}
\usage{
NeuroPath_Calc(GN, Path, Exp, MET)
}
\arguments{
\item{GN}{a character ENSG Gene name eg 'VAMP1|P23763'}

\item{Path}{the character string of Neuropath column to use for model eg. EITHER: 'ceradsc', 'braaksc', 'cogdx', 'dcfdx_lv'}

\item{MET}{Metadata data frame}

\item{EXP}{Expression data frame}
}
\value{
A vector
}
\description{
A function
}
