% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Analysis_Functions.R
\name{spvbsrBootstrap}
\alias{spvbsrBootstrap}
\title{A function}
\usage{
spvbsrBootstrap(y, x, nsamp = 100, cores = 8)
}
\arguments{
\item{y}{}

\item{x}{}

\item{nsamp}{a numeric}

\item{cores}{a numeric}
}
\value{
A matrix
}
\description{
A function
}
