% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Analysis_Functions.R
\name{TMT_Express_Load}
\alias{TMT_Express_Load}
\title{SynID <- 'syn23583548'
A function}
\usage{
TMT_Express_Load(SynID, names)
}
\arguments{
\item{SynID}{a synapse ID of a proteomics csv matrix eg. syn21266454 or syn21266454}

\item{names}{number corresponding to the row that are the rownames (Use 0 if none apply )}
}
\value{
The dataframe pulled from synapse
}
\description{
SynID <- 'syn23583548'
A function
}
